changed email sent option to admin site, also added email model so we know what emails we have send
