+++ apps/dishes/views.py
@@ -1,15 +1,10 @@
-from pprint import pprint
-from django.contrib.auth.decorators import login_required
-import string
-from unicodedata import name
 from django.shortcuts import redirect, render
 from django.contrib.auth.models import User
 from .models import Breeds, Puppy, ContactDetail, Menus
 from .forms import PuppyForm, MenusForm
 from .utils import *
 import json
-import os
-import re
+
 
 YOUNG_AGE = 2
 MIDDLE_AGE = 4
@@ -128,7 +123,7 @@ def formulate_home(request, menu_id=None):
 
 
 # @login_required(login_url='login')
-def menusHome(request, pk=None):
+def menus_home(request, pk=None):
     page = 'menus'
     available_menus = Menus.objects.all()
     puppy_data = None
@@ -195,7 +190,7 @@ def menusHome(request, pk=None):
     return render(request, 'dishes/menus.html', context)
 
 
-def editPet(request, pk):
+def edit_pet(request, pk):
     puppy = Puppy.objects.get(id=pk)
     form = PuppyForm(instance=puppy)
 
@@ -209,7 +204,7 @@ def editPet(request, pk):
     return render(request, 'dishes/home.html', context)
 
 
-def createMenu(request):
+def create_menu(request):
     form = MenusForm()
 
     if request.method == 'POST':
@@ -230,13 +225,13 @@ def createMenu(request):
     return render(request, 'menus/create-menu.html', context)
 
 
-def listMenus(request):
+def list_menus(request):
     menus = Menus.objects.all()
     context = {'menus': menus, 'page': 'menus-list'}
     return render(request, 'menus/menus-list.html', context)
 
 
-def deleteMenu(request, pk):
+def delete_menu(request, pk):
     menu = Menus.objects.get(id=pk)
 
     if request.method == 'POST':
@@ -245,12 +240,12 @@ def deleteMenu(request, pk):
         return redirect('menus-list')
 
 
-def updateMenu(request, pk):
+def update_menu(request, pk):
     menu = Menus.objects.get(id=pk)
 
     if request.method == 'POST':
         menu.delete()
-        return createMenu(request)
+        return create_menu(request)
 
     menu.percents = convert_json_to_string(menu.percents)
     menu.nutrition_information = convert_json_to_string(menu.nutrition_information)
@@ -261,7 +256,7 @@ def updateMenu(request, pk):
     return render(request, 'menus/create-menu.html', context)
 
 
-def menuSelection(request):
+def menu_selection(request):
     menus = Menus.objects.all()
 
     # get all puppies grams related to user
@@ -297,7 +292,7 @@ def menuSelection(request):
     return render(request, 'dishes/menu-selection.html', context)
 
 
-def menuDetail(request, menu_id, pet_id=None):
+def menu_detail(request, menu_id, pet_id=None):
     menu = Menus.objects.get(id=menu_id)
     menu.percents = json.loads(menu.percents)
     menu.nutrition_information = json.loads(menu.nutrition_information)
@@ -350,7 +345,7 @@ def menuDetail(request, menu_id, pet_id=None):
     return render(request, 'dishes/menu-selection.html', context)
 
 
-def menuPet(request, pk):
+def menu_pet(request, pk):
     puppy = Puppy.objects.get(id=pk)
 
     if puppy.menu is None:
