--- controls/templates/breed-detail.html
@@ -1,38 +1,0 @@
-{% extends 'main.html' %}
-{% block content %}
-
-<div class="main_container">
-    <div class="main_form_container">
-        <div class="wrapper">
-          <section class="form login">'
-            <div class="form_back_section">
-                <div class="back_icon_container">
-                    <i class="fa fa-angle-left"></i>
-                    <a href="{% url 'breeds' %}"><p>Volver</p></a>
-                </div>
-            </div>
-            <form method="post">
-                {% csrf_token %}
-                <div class="field input">
-                    <input type="text" name="name"  value="{{ breed.name }}">
-                </div>
-
-                <div class="field input">
-                    <input type="text" name="breed_size" value="{{ breed.breed_size }}">
-                </div>
-
-                <div class="field input">
-                    <input type="text" name="life_span"  value="{{ breed.life_span }}">
-                </div>
-                
-                <div class="field button">
-                    <input type="submit" value="submit" class="btn" />
-                </div>
-            </form>
-
-            </section>
-        </div>
-    </div>      
-</div>
-
-{% endblock %}
