--- controls/templates/breeds.html
@@ -1,35 +1,0 @@
-{% extends 'main.html' %}
-{% load static %}
-{% load humanize %}
-{% block content %}
-
-
-<div class="main_container">
-    <div class="main_form_container">
-        <div class="wrapper">
-          <section class="form login">
-                <table class="breed-table">
-                    <tr>
-                        <th>Name</th>
-                        <th>Size</th>
-                        <th>options</th>
-                    </tr>
-                    {% for breed in breeds %}
-                    <tr>
-                        <td>{{breed.name}}</td>
-                        <td>{{breed.breed_size}}</td>
-                        <td>
-                            <form method="POST">
-                                {% csrf_token %}
-                                <a href="{% url 'breed-detail' breed.id %}">Edit</a>
-                            </form>
-                        </td>
-                    </tr>
-                    {% endfor %}
-                </table>
-            </section>
-        </div>
-    </div>      
-</div>
-
-{% endblock %}
